+++ package.json
@@ -17,6 +17,7 @@
     "react": "^18.2.0",
     "react-calendar": "^4.6.1",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.11",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^6.19.0",
     "react-select": "^5.8.0",
