+++ src/components/Main/Constant/areaCategory.ts
@@ -1,0 +1,6 @@
+export const areaCategory = [
+  { area: "서울", areaCode: "SEOUL" },
+  { area: "경기", areaCode: "GYEONGGI" },
+  { area: "강원", areaCode: "GANGWON" },
+  { area: "부산", areaCode: "BUSAN" },
+];
