+++ src/pages/mainPage/MainPage.tsx
@@ -1,16 +1,12 @@
-import { useState } from "react";
+import { Suspense, useState } from "react";
 import Beloved from "../../components/Main/Beloved";
 import CategoryTab from "../../components/Main/CategoryTab";
 import MainCarousel from "../../components/Main/MainCarousel";
 import SearchBar from "../../components/Main/SearchBar";
 import styled from "styled-components";
-
-export const areaCategory = [
-  { area: "서울", areaCode: "SEOUL" },
-  { area: "경기", areaCode: "GYEONGGI" },
-  { area: "강원", areaCode: "GANGWON" },
-  { area: "부산", areaCode: "BUSAN" },
-];
+import SkeletonPage from "../../components/Main/SkeletonPage";
+import { ErrorBoundary } from "react-error-boundary";
+import ErrorPage from "../../components/Main/ErrorPage";
 
 export default function MainPage() {
   const [currentPage, setCurrentPage] = useState(0);
@@ -19,7 +15,11 @@ export default function MainPage() {
     <Main>
       <SearchBar />
       <CategoryTab setCurrentPage={setCurrentPage} />
-      <MainCarousel currentPage={currentPage} />
+      <ErrorBoundary FallbackComponent={ErrorPage} onReset={() => setCurrentPage(0)}>
+        <Suspense fallback={<SkeletonPage />}>
+          <MainCarousel currentPage={currentPage} />
+        </Suspense>
+      </ErrorBoundary>
       <Beloved />
     </Main>
   );
